# apps/tekimax/api_server/openapi.yaml
openapi: 3.0.0
info:
  title: Tekimax API Gateway
  x-logo:
    url: "/logo.png"
    altText: "Tekimax Logo"
    backgroundColor: "#FFFFFF"
  description: |
    # Tekimax API Gateway
    
    > **Sovereign AI Trust & Orchestration Layer**
    
    The Tekimax API serves as the central nervous system for the **Sovereign Engine**, orchestrating a secure L1-L4 pipeline:
    1.  **L1 (Conditioning)**: Privacy scrubbing & input isolation.
    2.  **L2 (Cognition)**: Neural inference and content generation.
    3.  **L3 (Authority)**: Cryptographic provenance and signing (C2PA).
    4.  **L4 (Compounding)**: Immutable audit logging.
    
    Every response is cryptographically signed with an `X-C2PA-Manifest` header, ensuring a verifiable chain of custody.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Gateway
security:
  - BearerAuth: []
...
paths:
...
  /api/tokenize:
    post:
      tags:
        - "Infrastructure"
      summary: Tokenize Text
      operationId: tokenize
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenizeRequest'
      responses:
        '200':
          description: Token list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenizeResponse'
  /api/detokenize:
    post:
      tags:
        - "Infrastructure"
      summary: Detokenize IDs
      operationId: detokenize
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetokenizeRequest'
      responses:
        '200':
          description: Reconstructed text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetokenizeResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: "Enter your key with the prefix: 'Bearer sk_...'"
  schemas:
...
    TokenizeRequest:
      type: object
      required: [model, prompt]
      properties:
        model: { type: string }
        prompt: { type: string }
      example:
        model: "gemini-3-flash-preview"
        prompt: "Hello world"
    DetokenizeRequest:
      type: object
      required: [model, tokens]
      properties:
        model: { type: string }
        tokens: { type: array, items: { type: integer } }
      example:
        model: "gemini-3-flash-preview"
        tokens: [15496, 5916]

    # --- Responses ---
...
    TokenizeResponse:
      type: object
      properties:
        tokens: { type: array, items: { type: integer } }
    DetokenizeResponse:
      type: object
      properties:
        prompt: { type: string }

    # --- Management Requests ---
...